requirements.txt
adspower==2.1.0
adspower-sdk==0.1.5
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.13.3
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
dateparser==1.2.1
defusedxml==0.7.1
distro==1.9.0
docutils==0.21.2
fake-useragent==2.0.3
Faker==35.2.0
filelock==3.17.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
id==1.5.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.2
jmespath==1.0.1
keyring==25.6.0
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
mysqlclient==2.2.7
nh3==0.2.21
oauth2client==4.1.3
openai==1.61.1
outcome==1.3.0.post0
packaging==24.2
parsel==1.10.0
pillow==11.2.1
playwright==1.51.0
pluggy==1.5.0
prometheus_client==0.22.1
Protego==0.4.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-wordpress-xmlrpc==2.3
pytz==2025.1
queuelib==1.7.0
readme_renderer==44.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rsa==4.9
schedule==1.2.2
Scrapy==2.12.0
scrapy-fake-useragent==1.4.4
scrapy-playwright==0.0.43
scrapyd-client==2.0.0
selenium==4.28.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
text-unidecode==1.3
tldextract==5.1.3
tqdm==4.67.1
transitions==0.9.2
trio==0.28.0
trio-websocket==0.11.1
twine==6.1.0
Twisted==24.11.0
typing_extensions==4.12.2
tzlocal==5.3.1
uberegg==0.1.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.3.0
w3lib==2.3.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
zope.interface==7.2
